refactor(react-kit): move auth and signing components into features directory
