+++ packages/react-kit/src/features/signing/SignatureRequest/SignatureRequest.test.tsx
@@ -1,7 +1,7 @@
 import { cleanup, fireEvent, render, screen } from '@testing-library/react'
 import { afterEach, describe, expect, it, vi } from 'vitest'
-import { createStore } from '../../store'
-import type { PendingRequest } from '../../types'
+import { createStore } from '../../../store'
+import type { PendingRequest } from '../../../types'
 
 // Must be a stable reference (like real wagmi) so the effect doesn't re-run on every render.
 const mockStore = createStore()
