+++ packages/react-kit/src/features/signing/SignatureRequest/index.tsx
@@ -1,7 +1,7 @@
 'use client'
 
-import { usePendingRequest } from '../../hooks/usePendingRequest.js'
-import { cn } from '../../shared/utils/common.js'
+import { usePendingRequest } from '../../../hooks/usePendingRequest.js'
+import { cn } from '../../../shared/utils/common.js'
 
 // todo: proper UI
 export function SignatureRequest() {
